Alteração de funcionalidades

+++ melhorias/src/main/java/com/cottonstar/melhorias/repository/ComentarioRepository.java
@@ -3,5 +3,5 @@ package com.cottonstar.melhorias.repository;
 import com.cottonstar.melhorias.model.Comentario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ComentarioRepository extends JpaRepository<Comentario, Integer> {
+public interface ComentarioRepository extends JpaRepository<Comentario, Long> {
 }
